ChatApp/project/requirements.txt
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
Brlapi==0.6.6
certifi==2019.11.28
cffi==1.13.2
channels==0.17.3
chardet==3.0.4
CodeConvert==3.0.2
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
daphne==0.15.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
django-admin==2.0.0
django-excel-response2==2.0.8
django-filter==2.2.0
django-models-ext==1.1.9
django-short-url==1.1.4
django-six==1.0.4
django-templated-mail==1.1.1
django-widget-tweaks==1.4.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
furl==2.1.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
isort==4.3.21
jwt==0.6.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
mccabe==0.6.1
mod-wsgi-httpd==2.4.41.1
more-itertools==8.0.2
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
orderedmultidict==1.0.1
packaging==19.2
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
pluggy==0.13.1
protobuf==3.0.0
psycopg2==2.8.4
py==1.8.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyee==6.0.0
Pygments==2.5.2
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.6.1
pylint==2.4.4
pymacaroons==0.13.0
pymitter==0.2.3
PyNaCl==1.1.2
pyparsing==2.4.5
pyRFC3339==1.0
pytest==5.3.2
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-memcached==1.59
pytz==2017.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
rope==0.14.0
screen==1.0.1
SecretStorage==2.3.1
shortuuid==0.5.0
simplejson==3.13.2
six==1.13.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
TimeConvert==1.5.1
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
validate-email==1.3
virtualenv==16.7.9
virtualenv-clone==0.5.3
wadllib==1.3.2
wcwidth==0.1.7
wrapt==1.11.2
xkit==0.0.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.7.1
